+++ test/config-in-subdir/subdir/playwright.config.ts
@@ -1,11 +1,12 @@
 import { defineConfig } from '@playwright/test';
-import { defineBddConfig } from 'playwright-bdd';
+import { defineBddConfig, cucumberReporter } from 'playwright-bdd';
 
 const testDir = defineBddConfig({
   paths: ['.'],
-  require: ['*.ts'],
+  require: ['steps.ts'],
 });
 
 export default defineConfig({
   testDir,
+  reporter: [cucumberReporter('html', { outputFile: 'actual-reports/report.html' })],
 });